third_party/blink/web_tests/platform/mac-mac14-arm64/fast/webgl/renderer-and-vendor-strings-expected.txt

CONSOLE ERROR: Uncaught Unable to fetch WebGL rendering context for Canvas
Verifies the contents of the RENDERER and VENDOR strings to avoid regressions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
